fix: Fixed duplicate canteen behavior
The canteen was giving duplicate penalties due to game engine bug where when item is updated, it is ticked again and thus the onItemUseFinish runs again. This commit fixed it by introducing a cooldown equal to the max use duration. Thus, it won't impact normal operation, but will filter out the duplicate operations.

+++ src/main/java/enginecrafter77/survivalinc/stats/impl/HydrationModifier.java
@@ -116,7 +116,7 @@ public class HydrationModifier implements StatProvider<SimpleStatRecord> {
 			}
 			else
 			{
-				WaterVolume verifiedvolume = WaterVolume.fromBlock(player.world, water_rt.getBlockPos(), (float)ModConfig.HYDRATION.sipVolume);
+				WaterVolume verifiedvolume = WaterVolume.fromBlock(player.world, water_rt.getBlockPos(), ModConfig.HYDRATION.sipVolume);
 				if(!volume.equals(verifiedvolume))
 				{
 					SurvivalInc.logger.warn("Server raytrace results not consistent with client's claim. Fabricated packet from {}?", ip);
@@ -267,7 +267,7 @@ public class HydrationModifier implements StatProvider<SimpleStatRecord> {
 			if(hydration == null)
 				return null;
 
-			WaterVolume volume = WaterVolume.fromBlock(player.world, water_rt.getBlockPos(), (float)ModConfig.HYDRATION.sipVolume);
+			WaterVolume volume = WaterVolume.fromBlock(player.world, water_rt.getBlockPos(), ModConfig.HYDRATION.sipVolume);
 
 			// Test whether WaterVolume#fromBlock returned actual WaterVolume instance
 			if(volume == null)
